Move extension install script files to __tests__
Move the extension install script files to the `__tests__` sub directory
like we do for all our other test files. This is only a rename, no other
changes were make.
